Moved log out button and added version info to SettingsActivity
