+++ app/src/main/java/com/frisky/icebreaker/SettingsActivity.java
@@ -1,18 +1,26 @@
 package com.frisky.icebreaker;
 
+import android.content.Intent;
+import android.content.pm.PackageInfo;
 import android.graphics.Typeface;
 import android.support.v4.content.res.ResourcesCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.TypedValue;
 import android.view.View;
+import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
+
+import com.google.firebase.auth.FirebaseAuth;
 
 public class SettingsActivity extends AppCompatActivity {
 
     ImageButton mBackButton;
     TextView mToolbarText;
+    Button mLogoutButton;
+
+    TextView mVersionText;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -34,5 +42,17 @@ public class SettingsActivity extends AppCompatActivity {
         Typeface typeface = ResourcesCompat.getFont(getApplicationContext(), R.font.museosans700);
         mToolbarText.setTypeface(typeface);
 
+        mLogoutButton = findViewById(R.id.button_logout);
+        mLogoutButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                FirebaseAuth.getInstance().signOut();
+                startActivity(new Intent(getApplicationContext(), LoginActivity.class));
+            }
+        });
+
+        mVersionText = findViewById(R.id.text_version);
+        String version = BuildConfig.VERSION_NAME;
+        mVersionText.setText(version);
     }
 }
