+++ app/src/main/java/com/frisky/icebreaker/profile/EditImagesAdapter.java
@@ -21,7 +21,7 @@ public class EditImagesAdapter extends RecyclerView.Adapter<EditImagesAdapter.Im
     private Context mContext;
     private FragmentActivity mActivity;
     private PickImageDialog pickImageDialog;
-    
+
     // Provide a reference to the views for each data item
     // Complex data items may need more than one view per item, and
     // you provide access to all the views for a data item in a view holder
