+++ app/src/main/res/values/strings.xml
@@ -5,6 +5,11 @@
     <string name="password">Password</string>
     <string name="comma">,</string>
     <string name="email">Email</string>
+    <string name="logout">Logout</string>
+    <string name="version">Version</string>
+    <string name="facebook">Facebook</string>
+    <string name="apply">Apply</string>
+    <string name="cancel">Cancel</string>
 
     <!--Login Page-->
     <string name="login">Login</string>
@@ -82,11 +87,6 @@
     <string name="pick_image">Pick Image</string>
     <string name="image_gallery">Choose from Gallery</string>
     <string name="image_camera">Take a new photo</string>
-
-    <!--App Common-->
-    <string name="facebook">Facebook</string>
-    <string name="apply">Apply</string>
-    <string name="cancel">Cancel</string>
 
     <string name="placeholder">dummy</string>
     <string name="placeholder_number">23</string>
